Convert `#find_txn` to work with new BloomNet Center
