+++ bloom_net_center_client.gemspec
@@ -32,7 +32,7 @@ Gem::Specification.new do |spec|
   spec.add_dependency "virtus", "~> 1.0"
   spec.add_dependency "httparty", "~> 0.14.0"
   spec.add_dependency "activesupport", ">= 3.0"
-  spec.add_dependency "api_client_base", "~> 0.2.0"
+  spec.add_dependency "api_client_base", "~> 1.2"
   spec.add_dependency "typhoeus", "~> 1.1"
   spec.add_development_dependency "bundler", "~> 1.13"
   spec.add_development_dependency "rake", "~> 10.0"
