+++ lib/bloom_net_center_client.rb
@@ -44,5 +44,6 @@ require "bloom_net_center_client/requests/create_txn_request"
 require "bloom_net_center_client/requests/find_txn_request"
 require "bloom_net_center_client/responses/base_response"
 require "bloom_net_center_client/responses/create_claim_response"
+require "bloom_net_center_client/responses/txn_response"
 require "bloom_net_center_client/responses/create_txn_response"
 require "bloom_net_center_client/responses/find_txn_response"
