+++ lib/bloom_net_center_client/requests/base_request.rb
@@ -25,5 +25,9 @@ module BloomNetCenterClient
       OpenSSL::HMAC.hexdigest("SHA256", secret, body)
     end
 
+    def default_body
+      {}.to_json
+    end
+
   end
 end
