+++ lib/bloom_net_center_client/responses/find_txn_response.rb
@@ -1,22 +1,7 @@
 module BloomNetCenterClient
-  class FindTxnResponse < BaseResponse
-
-    attribute :txn, Object, lazy: true, default: :default_txn
+  class FindTxnResponse < TxnResponse
 
     private
-
-    def default_txn
-      txn_data = body[:data].first
-      return nil if txn_data.nil?
-      txn_attrs = txn_data.slice(*%i(
-        id
-        recipient_first_name
-        recipient_last_name
-        ref_no
-        status
-      ))
-      BloomNetCenterClient::Txn.new(txn_attrs)
-    end
 
     def default_success
       code == 200 && txn.present?
